refactor(agent): agent 目录域合并为 agent-files.ts
把同属「on-disk agent 文件」职责的四个小模块——layout(布局解析) / templates(初始化模版) /
scaffold(脚手架) / load(现读装配上下文+规则)——合并为单一领域模块 agent-files.ts，按
布局 → 模版 → 脚手架 → 读取 的次序收口。记忆**写入**侧（memory.ts 的 appendAgentNotes）按读/写
职责分明保留独立。

- 删除 layout.ts / templates.ts / scaffold.ts / load.ts，内容并入 agent-files.ts。
- index 公共导出改走 agent-files；memory.ts 的 resolveAgentPaths 引用改路径。外部消费方均经
  @meebox/agent 包根导入，公共 API 不变。
- 顺带：scaffoldAgentDir 的空参异常按「包内异常统一英语」规约改为英文。
- 测试 load.test.ts 重命名为 agent-files.test.ts。

纯结构调整、行为不变；agent 49 tests / lint / typecheck / desktop build 全绿。
