+++ packages/agent/src/index.ts
@@ -1,12 +1,14 @@
 export { AGENT_FILES, AGENT_RULES_SUBDIR, READ_TOOLS, MUTATING_TOOLS } from './constants.js';
-export { resolveAgentPaths } from './layout.js';
-export type { AgentContextKind } from './layout.js';
-export { loadAgentContext, loadAgentRules } from './load.js';
-export { scaffoldAgentDir } from './scaffold.js';
+export {
+  resolveAgentPaths,
+  loadAgentContext,
+  loadAgentRules,
+  scaffoldAgentDir,
+  AGENT_TEMPLATES,
+} from './agent-files.js';
+export type { AgentContextKind, AgentTemplate } from './agent-files.js';
 export { appendAgentNotes } from './memory.js';
 export type { MemoryNote, WritableAgentFile } from './memory.js';
-export { AGENT_TEMPLATES } from './templates.js';
-export type { AgentTemplate } from './templates.js';
 export type { AgentContext, AgentContextFiles, LoadAgentContextOptions } from './types.js';
 export { assembleSystemContext } from './assemble.js';
 export type {
